pyproject.toml
[project]
name = "fichub-net"
requires-python = ">=3.14"
version = "0.1.0"
dependencies = [
	"Flask",
	"requests",
	"psycopg2",
	"python-dateutil",
	"EbookLib @ git+https://github.com/FicHub/ebooklib@ef040aa",
	"oil @ git+https://github.com/FanFicDev/python-oil@b575c0c",
	"elasticsearch==8.19.3",
	"psutil",
]

[dependency-groups]
dev = [
	"hatch",
	"mypy",
	"pytest",
	"pytest-cov",
	"responses",
	"ruff",
	"testcontainers",
	"tox",
	"tox-uv",
	"types-psutil",
	"types-psycopg2",
	"types-python-dateutil",
	"types-requests",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/fichub_net"]

[tool.pytest.ini_options]
markers = [
	"slow: marks tests as slow (deselect with '-m \"not slow\"')",
	"elasticsearch: mark tests that use elasticsearch (deselect with '-m \"not elasticsearch\"')",
]
pythonpath = ["src/"]

[tool.coverage.run]
omit = ["tests/*"]

[tool.coverage.report]
exclude_also = [
	"if __name__ == .__main__.:",
	"if typing.TYPE_CHECKING:",
	"if TYPE_CHECKING:",
]

[tool.mypy]
follow_imports = "normal"
ignore_missing_imports = false
python_version = "3.14"

# https://mypy.readthedocs.io/en/stable/config_file.html#disallow-dynamic-typing
disallow_any_unimported = true
#disallow_any_expr = true
#disallow_any_decorated = true
#disallow_any_explicit = true
disallow_any_generics = true
disallow_subclassing_any = true

# https://mypy.readthedocs.io/en/stable/config_file.html#untyped-definitions-and-calls
#disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true

# https://mypy.readthedocs.io/en/stable/config_file.html#none-and-optional-handling
# default: implicit_optional = false
# default: strict_optional = true

# https://mypy.readthedocs.io/en/stable/config_file.html#configuring-warnings
warn_redundant_casts = true
warn_unused_ignores = true
# default: warn_no_return = true
warn_return_any = true
warn_unreachable = true

# https://mypy.readthedocs.io/en/latest/config_file.html#miscellaneous-strictness-flags
#implicit_reexport = false

# https://mypy.readthedocs.io/en/stable/config_file.html#configuring-error-messages
show_error_context = true
show_column_numbers = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "testcontainers.*"
ignore_missing_imports = true

[tool.ruff]
exclude = ["dev-docker-compose"]

[tool.ruff.lint]
extend-select = [
	"F",     # pyflakes
	# "E",     # pycodestyle error
	"W",     # pycodestyle warning
	# "C90",   # mccabe
	"I",     # isort
	"N",     # pep8-naming
	# "D",     # pydocstyle
	"UP",    # pyupgrade
	"YTT",   # flake8-2020
	# "ANN",   # flake8-annotations
	"ASYNC", # flake8-async
	# "S",     # flake8-bandit
	# "BLE",   # flake8-blind-except
	# "FBT",   # flake8-boolean-trap
	"B",     # flake8-bugbear
	"A",     # flake8-builtins
	# "COM",   # flake8-commas -- affects formatting
	# "CPY",   # flake8-copyright -- preview
	"C4",    # flake8-comprehensions
	"DTZ",   # flake8-datetimez
	"T10",   # flake8-debugger
	"DJ",    # flake8-django
	"EM",    # flake8-errmsg
	"EXE",   # flake8-executable
	# "FA",    # flake8-future-annotations -- should upgrade python
	"ISC",   # flake8-implicit-str-concat
	"ICN",   # flake8-import-conventions
	"LOG",   # flake8-logging
	# "G",     # flake8-logging-format
	"INP",   # flake8-no-pep420
	"PIE",   # flake8-pie
	# "T20",   # flake8-print -- switch to logging
	"PYI",   # flake8-pyi
	"PT",    # flake8-pytest-style
	"Q",     # flake8-quotes
	"RSE",   # flake8-raise
	"RET",   # flake8-return
	"SLF",   # flake8-self
	"SLOT",  # flake8-slots
	"SIM",   # flake8-simplify
	"TID",   # flake8-tidy-imports
	"TCH",   # flake8-type-checking
	"INT",   # flake8-gettext
	"ARG",   # flake8-unused-arguments
	"PTH",   # flake8-use-pathlib
	"TD",    # flake8-todos
	"FIX",   # flake8-fixme
	"ERA",   # eradicate
	"PD",    # pandas-vet
	"PGH",   # pygrep-hooks
	"PL",    # Pylint
	"TRY",   # tryceratops
	"FLY",   # flynt
	"NPY",   # NumPy-specific rules
	"AIR",   # Airflow
	"PERF",  # Perflint
	"FURB",  # refurb
	"RUF",   # Ruff-specific rules
]

ignore = [
	"ISC001",  # single-line-implicit-string-concatenation
	"TD002",   # missing-todo-author
	"TD003",   # missing-todo-link
	"FIX002",  # line-contains-todo
	"PLR0911", # too-many-return-statements
	"PLR0912", # too-many-branches
	"PLR0913", # too-many-arguments
	"PLR0915", # too-many-statements
	"TRY300",  # try-consider-else
	"N999",    # invalid-module-name
	"EXE003",  # shebang-missing-python
]

# "ANN101", "ANN102", "COM812"

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
	"S101",   # assert
	"INP001", # implicit-namespace-package
]

#"tests/*" = ["PLR2004"]

[tool.ruff.lint.isort]
force-to-top = ["typing"]
force-sort-within-sections = true

[tool.tox]
legacy_tox_ini = """
[tox]
isolated_build = True
env_list =
  type, lint, py3.{12,11,10}
skip_missing_interpreters = true

[testenv]
commands = pytest -v --tb=short --basetemp={envtmpdir} {posargs}

[testenv:type]
skip_install = true
commands = mypy {posargs:src tests}

[testenv:lint]
skip_install = true
commands =
  ruff format {posargs:src tests}
  ruff check {posargs:src tests}
"""
